feat: Add assign method to XORList
Add two new methods `assign` to the XORList class:
- `assign(InputIt first, InputIt last)` for assigning values from a range
- `assign(std::initializer_list<T> ilist)` for assigning values from an initializer list

These methods clear the current list and then populate it with elements from either a range or an initializer list. The purpose is to provide flexibility in how users can assign values to the XORList.

Tests have been added for both scenarios:
1. Assigning values from a range using `AssignRange` test case
2. Assigning values from an initializer list using `AssignInitializerList` test case

The tests ensure that after calling these methods, the XORList contains the expected elements in the correct order.